requirements.txt
binary==1.0.2
mysql-connector-python==9.3.0
numpy==2.3.1
pandas==2.3.1
psycopg==3.2.9
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tzdata==2025.2
